Rename file
https://coveord.atlassian.net/browse/JSUI-1954
